+++ wayland-backend/src/test/object_args.rs
@@ -68,7 +68,14 @@ macro_rules! impl_server_objectdata {
                 None
             }
 
-            fn destroyed(&self, _: &mut (), _: $server_backend::ClientId, _: $server_backend::ObjectId) {}
+            fn destroyed(
+                self: Arc<Self>,
+                _: &$server_backend::Handle,
+                _: &mut (),
+                _: $server_backend::ClientId,
+                _: $server_backend::ObjectId
+            ) {
+            }
         }
 
         impl $server_backend::GlobalHandler<()> for ServerData {