+++ wayland-backend/src/test/protocol_error.rs
@@ -223,7 +223,14 @@ impl<D> server_rs::ObjectData<D> for ProtocolErrorServerData {
         None
     }
 
-    fn destroyed(&self, _: &mut D, _: server_rs::ClientId, _: server_rs::ObjectId) {}
+    fn destroyed(
+        self: Arc<Self>,
+        _handle: &server_rs::Handle,
+        _: &mut D,
+        _: server_rs::ClientId,
+        _: server_rs::ObjectId,
+    ) {
+    }
 }
 
 impl<D> server_sys::ObjectData<D> for ProtocolErrorServerData {
@@ -238,7 +245,14 @@ impl<D> server_sys::ObjectData<D> for ProtocolErrorServerData {
         None
     }
 
-    fn destroyed(&self, _: &mut D, _: server_sys::ClientId, _: server_sys::ObjectId) {}
+    fn destroyed(
+        self: Arc<Self>,
+        _handle: &server_sys::Handle,
+        _: &mut D,
+        _: server_sys::ClientId,
+        _: server_sys::ObjectId,
+    ) {
+    }
 }
 
 expand_test!(protocol_error_in_request_without_object_init, {